Added Phone field to to ApplicationUser model. Separated Manage ViewModels into separate class files.
