+++ Vidly/ViewModels/ConfigureTwoFactorViewModel.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Vidly.ViewModels
+{
+    public class ConfigureTwoFactorViewModel
+    {
+        public string SelectedProvider { get; set; }
+        public ICollection<System.Web.Mvc.SelectListItem> Providers { get; set; }
+    }
+}
