+++ Vidly/ViewModels/FactorViewModel.cs
@@ -1,0 +1,7 @@
+﻿namespace Vidly.ViewModels
+{
+    public class FactorViewModel
+    {
+        public string Purpose { get; set; }
+    }
+}
