+++ Vidly/ViewModels/VerifyPhoneNumberViewModel.cs
@@ -1,0 +1,17 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Vidly.ViewModels
+{
+
+    public class VerifyPhoneNumberViewModel
+    {
+        [Required]
+        [Display(Name = "Code")]
+        public string Code { get; set; }
+
+        [Required]
+        [Phone]
+        [Display(Name = "Phone Number")]
+        public string PhoneNumber { get; set; }
+    }
+}
