+++ Vidly/Views/Account/ExternalLoginConfirmation.cshtml
@@ -25,9 +25,16 @@
         </div>
     </div>
     <div class="form-group">
-        @Html.LabelFor(m => m.Email, new { @class = "col-md-2 control-label" })
+        @Html.LabelFor(m => m.DrivingLicense, new { @class = "col-md-2 control-label" })
         <div class="col-md-10">
             @Html.TextBoxFor(m => m.DrivingLicense, new { @class = "form-control" })
+            @Html.ValidationMessageFor(m => m.DrivingLicense, "", new { @class = "text-danger" })
+        </div>
+    </div>
+    <div class="form-group">
+        @Html.LabelFor(m => m.Phone, new { @class = "col-md-2 control-label" })
+        <div class="col-md-10">
+            @Html.TextBoxFor(m => m.Phone, new { @class = "form-control" })
             @Html.ValidationMessageFor(m => m.DrivingLicense, "", new { @class = "text-danger" })
         </div>
     </div>
